+++ pg_documentdb/src/aggregation/bson_aggregation_nested_pipeline.c
@@ -55,7 +55,6 @@
 
 const int MaximumLookupPipelineDepth = 20;
 extern bool EnableLookupIdJoinOptimizationOnCollation;
-extern bool EnableNowSystemVariable;
 extern bool EnableLookupInnerJoin;
 extern bool EnableOperatorVariablesInLookup;
 extern bool EnableUseForeignKeyLookupInline;
@@ -2026,7 +2025,7 @@ OptimizeLookup(LookupArgs *lookupArgs,
 	if (lookupArgs->let)
 	{
 		/* Validate the lookupArgs->let if the left query had no let variables specified in it. */
-		/* With EnableNowSystemVariable, time system variables are stored in the left query's variableSpec even in the absence of let. */
+		/* Time system variables are stored in the left query's variableSpec even in the absence of let. */
 		/* Thus, if the left query's variableSpec contains only the time system variables, */
 		/* we perform a validation on the lookupArgs->let. */
 		if (leftQueryContext->variableSpec == NULL)
@@ -2040,8 +2039,7 @@ OptimizeLookup(LookupArgs *lookupArgs,
 
 			ParseVariableSpec(&varsValue, nullContext, &parseContext);
 		}
-		else if (EnableNowSystemVariable &&
-				 IsA(leftQueryContext->variableSpec, Const))
+		else if (IsA(leftQueryContext->variableSpec, Const))
 		{
 			Node *specNode = (Node *) leftQueryContext->variableSpec;
 			Const *specConst = (Const *) specNode;