+++ pg_documentdb/src/commands/delete.c
@@ -405,9 +405,7 @@ BuildBatchDeletionSpec(bson_iter_t *deleteCommandIter, pgbsonsequence *deleteDoc
 	if (EnableVariablesSupportForWriteCommands)
 	{
 		TimeSystemVariables *timeSysVars = NULL;
-		bool isWriteCommand = true;
-		pgbson *parsedVariables = ParseAndGetTopLevelVariableSpec(&let, timeSysVars,
-																  isWriteCommand);
+		pgbson *parsedVariables = ParseAndGetTopLevelVariableSpec(&let, timeSysVars);
 		batchSpec->variableSpec = ConvertPgbsonToBsonValue(parsedVariables);
 	}